+++ environment/src/main/java/jetbrains/exodus/tree/btree/InternalPageMutable.java
@@ -97,7 +97,7 @@ public class InternalPageMutable extends BasePageMutable {
     }
 
     @Override
-    protected int getType() {
+    protected byte getType() {
         return tree.getInternalPageType();
     }
